Enable the idle timeout by adjust RX FIFO threshold and idle timeout time
If a device must support reading input DMX packets of variable length,
it needs to call dmx_recieve with the max DMX universe size of 513
bytes. If it is receiving packets smaller than 513, then it will not
signal a new DMX packet to the waiting task until the break of the
following packet. In testing, this was seen to miss every other packet.

By enabling the idle interrupt, we can now signal to the waiting task
that we've received a packet < 513 almost directly after receiving it
instead of waiting until the next packet.

By setting an RX FIFO threshold value of a multiple of the max 513
bytes, this ensures that the standard logic of receiving a packet will
still fire immediately at the end of the 513 byte packets without having
to wait for the idle interrupt.

Since the RX FIFO of the esp32 cannot be set above 128, the closest
multiple of 513 is 57. This will result in the one small edge case for
packet sizes that are multiples of 57, but not the full 513 (57, 114,
etc). In these cases, the DMX parsing logic will not signal a new packet
until the break starts for the following packet (which is the current
behavior for _all_ packets less than the byte number passed to
dmx_receive). This is an acceptable edge case since it's so unlikely to
happen, and causes no degradation of the current behavior if it does.